properly utilize keys import for test suite

+++ tests/test.sh
@@ -9,6 +9,7 @@
 # REQUIREMENTS
 
 # eris installed locally
+# eris-keys installed locally
 
 # ----------------------------------------------------------
 # USAGE
@@ -28,6 +29,8 @@ else
   repo=$GOPATH/src/$base
   circle=false
 fi
+branch=${CIRCLE_BRANCH:=master}
+branch=${branch/-/_}
 
 # Define now the tool tests within the Docker container will be booted from docker run
 entrypoint="/home/eris/test_tool.sh"
@@ -36,9 +39,11 @@ testuser=eris
 
 # Other needed variables
 uuid=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 12 | head -n 1)
-was_running=0
 key1="RqsMlojbh9RwUUVTfIXvLgBq4PzYxG3NNeD1eUWCUxyC2trhU0BkEoM/gJhDntS13MJsC0PCd17/6LkRtJ0Bgg=="
+key1=`echo -n "$key1" | base64 -d | hexdump -ve '1/1 "%.2X"'`
 key2="cyAlnxHeURCJgNSPNea3aTjmSId0MfKykAR/iSRN19132APZNKM1FETmvHV83w60ds4PVvl153a+4dtqCC4q+Q=="
+key2=`echo -n "$key2" | base64 -d | hexdump -ve '1/1 "%.2X"'`
+was_running=0
 
 # ---------------------------------------------------------------------------
 # Needed functionality
@@ -77,34 +82,52 @@ echo "Getting Setup"
 if [ "$circle" = true ]
 then
   export ERIS_PULL_APPROVE="true"
-  eris init --yes --skip-pull
+  eris init --yes --skip-pull 1>/dev/null
+  echo 'ports = [ "4767:4767" ]' >> ~/.eris/services/keys.toml
 fi
 
 is_it_running keys
-# mkdir ~/.eris/data/keys/data
-# cp $GOPATH/src/github.com/eris-ltd/eris-pm/tests/fixtures/keys/* ~/.eris/data/keys/keys/data/.
-# eris data import keys
 if [ $? -eq 0 ]
 then
-  eris services start keys
+  eris services start keys 1>/dev/null
+  keysHost=$(eris services inspect keys NetworkSettings.IPAddress)
+  eris-keys import "$key1" --no-pass --host $keysHost 1>/dev/null
+  eris-keys import "$key2" --no-pass --host $keysHost 1>/dev/null
+else
+  eris-keys import "$key1" --no-pass 1>/dev/null
+  eris-keys import "$key2" --no-pass 1>/dev/null
 fi
-eris chains new epm-tests-$uuid --dir tests/fixtures/chaindata
+
+eris chains new epm-tests-$uuid --dir tests/fixtures/chaindata 1>/dev/null
 sleep 5
 echo "Setup complete"
+
 echo ""
-docker run --rm --link eris_service_keys_1:keys --link eris_chain_epm-tests-"$uuid"_1:chain --entrypoint $entrypoint --user $testuser --env CHAINID=epm-tests-$uuid $testimage
+if [ "$circle" = true ]
+then
+  if [[ "$branch" = "master" ]]
+  then
+    branch="latest"
+  fi
+  docker run --link eris_service_keys_1:keys --link eris_chain_epm-tests-"$uuid"_1:chain --entrypoint $entrypoint --user $testuser --env CHAINID=epm-tests-$uuid $testimage:$branch
+else
+  docker run --rm --link eris_service_keys_1:keys --link eris_chain_epm-tests-"$uuid"_1:chain --entrypoint $entrypoint --user $testuser --env CHAINID=epm-tests-$uuid $testimage
+fi
 test_exit=$?
 
 # ---------------------------------------------------------------------------
 # Cleaning up
 
 echo ""
-eris chains stop -xf epm-tests-$uuid
-eris chains rm -xof epm-tests-$uuid
-rm -rf ~/.eris/data/epm-tests-*
-if [ "$was_running" -eq 0 ]
+if [ "$circle" = false ]
 then
-  eris services stop keys
+  eris chains stop -xf epm-tests-$uuid
+  eris chains rm -xf epm-tests-$uuid
+  rm -rf ~/.eris/data/epm-tests-*
+  if [ "$was_running" -eq 0 ]
+  then
+    eris services stop -rx keys
+  fi
 fi
 if [ "$test_exit" -eq 0 ]
 then
